+++ server/lib/src/router/api/api.dart
@@ -9,6 +9,7 @@ import 'challenge/challenge.dart';
 import 'game/game.dart';
 import 'games/games.dart';
 import 'player/player.dart';
+import 'players/players.dart';
 import 'settings/settings.dart';
 
 /// Router that handles requests on `/ICServerRouter.api`.
@@ -16,8 +17,11 @@ class ApiRouter implements RouterInterface {
   /// Router that handles auth requests.
   final AuthRouter _authRouter;
 
-  /// Router that handles requests regarding players.
+  /// Router that handles requests for single player.
   final PlayerRouter _playerRouter;
+
+  /// Router that handles requests regarding players.
+  final PlayersRouter _playersRouter;
 
   /// Router that handles requests for games.
   final ChallengeRouter _challengeRouter;
@@ -38,12 +42,14 @@ class ApiRouter implements RouterInterface {
   ApiRouter({
     AuthRouter? authRouter,
     PlayerRouter? playerRouter,
+    PlayersRouter? playersRouter,
     ChallengeRouter? challengeRouter,
     SettingsRouter? settingsRouter,
     GameRouter? gameRouter,
     GamesRouter? gamesRouter,
   })  : _authRouter = authRouter ?? AuthRouter(),
         _playerRouter = playerRouter ?? PlayerRouter(),
+        _playersRouter = playersRouter ?? PlayersRouter(),
         _challengeRouter = challengeRouter ?? ChallengeRouter(),
         _settingsRouter = settingsRouter ?? SettingsRouter(),
         _gameRouter = gameRouter ?? const GameRouter(),
@@ -63,6 +69,8 @@ class ApiRouter implements RouterInterface {
         return await _authRouter.handle(request);
       case ICServerRoute.apiPlayer:
         return await _playerRouter.handle(request);
+      case ICServerRoute.apiPlayers:
+        return await _playersRouter.handle(request);
       case ICServerRoute.apiChallenge:
         return await _challengeRouter.handle(request);
       case ICServerRoute.apiSettings: